+++ src/main/java/org/aarboard/nextcloud/api/provisioning/Quota.java
@@ -1,9 +1,12 @@
 package org.aarboard.nextcloud.api.provisioning;
 
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 
 @XmlAccessorType(XmlAccessType.FIELD)
+@JsonIgnoreProperties(ignoreUnknown = true)
 public class Quota {
     private long free;
     private long used;