+++ src/test/java/org/aarboard/nextcloud/api/ATestClass.java
@@ -25,10 +25,14 @@ import org.junit.Before;
  */
 public class ATestClass {
     public static final String TESTUSER = "testuser";
+    public static final String TESTUSER_WITH_INVALID_CHAR = "test@user.org";
     public static final String TESTGROUP = "testgroup";
     public static final String TEST_FOLDER = "new-test-folder";
+    public static final String TEST_FOLDER_RENAMED = "new-test-folder-renamed";
     public static final String TESTFILE1 = "test1.txt";
+    public static final String TESTFILE1_RENAMED = "test1-renamed.txt";
     public static final String TESTFILE2 = "test2.txt";
+    public static final String TESTFILE2_RENAMED = "test2-renamed.txt";
     public static final String TESTFILE3 = "test3-äöüÖÄÜ and with plus + sign \u32A2.txt";
     public static final String TESTFILE4 = "test4.txt";
     public static final String TESTFILE6 = "test6.txt";