+++ src/test/java/org/aarboard/nextcloud/api/TestUserGroupAdmin.java
@@ -16,23 +16,16 @@
  */
 package org.aarboard.nextcloud.api;
 
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertFalse;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
-
-import java.io.ByteArrayInputStream;
-import java.io.InputStream;
-import java.util.Collection;
-import java.util.List;
-
 import org.aarboard.nextcloud.api.provisioning.User;
 import org.aarboard.nextcloud.api.provisioning.UserData;
-import org.junit.AfterClass;
-import org.junit.Before;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.runners.MethodSorters;
+
+import java.util.Collection;
+import java.util.List;
+
+import static org.junit.Assert.*;
 
 /**
  *
@@ -52,6 +45,16 @@ public class TestUserGroupAdmin extends ATestClass {
     }
 
     @Test
+    public void t01_01_testCreateUserWithInvalidChar() {
+        System.out.println("createUserWithInvalidChar");
+        if (_nc != null)
+        {
+            boolean result = _nc.createUser(TESTUSER_WITH_INVALID_CHAR, "aBcDeFg123456");
+            assertTrue(result);
+        }
+    }
+
+    @Test
     public void t02_testGetUsers() {
         System.out.println("getUsers");
         if (_nc != null)
@@ -59,6 +62,18 @@ public class TestUserGroupAdmin extends ATestClass {
             Collection<String> result = _nc.getUsers();
             assertNotNull(result);
             assertTrue(result.contains(TESTUSER));
+        }
+    }
+
+    @Test
+    public void t02_01_testGetUsersDetails() {
+        System.out.println("getUsersDetails");
+        if (_nc != null)
+        {
+            Collection<User> result = _nc.getUsersDetails();
+            assertNotNull(result);
+            assertTrue(result.stream().anyMatch(user -> TESTUSER.equals(user.getId())));
+            assertTrue(result.stream().anyMatch(user -> TESTUSER_WITH_INVALID_CHAR.equals(user.getId())));
         }
     }
 
@@ -73,6 +88,22 @@ public class TestUserGroupAdmin extends ATestClass {
             Collection<String> result = _nc.getUsers(search, limit, offset);
             assertNotNull(result);
             assertTrue(result.contains(TESTUSER));
+        }
+    }
+
+    @Test
+    public void t03_01_testGetUsersDetails_3args() {
+        System.out.println("getUsersDetails");
+        if (_nc != null)
+        {
+            int limit = 1;
+            int offset = -1;
+            Collection<User> result1 = _nc.getUsersDetails(TESTUSER, limit, offset);
+            assertNotNull(result1);
+            assertTrue(result1.stream().anyMatch(user -> TESTUSER.equals(user.getId())));
+            Collection<User> result2 = _nc.getUsersDetails(TESTUSER_WITH_INVALID_CHAR, limit, offset);
+            assertNotNull(result2);
+            assertTrue(result2.stream().anyMatch(user -> TESTUSER_WITH_INVALID_CHAR.equals(user.getId())));
         }
     }
 
@@ -245,6 +276,16 @@ public class TestUserGroupAdmin extends ATestClass {
         if (_nc != null)
         {
             boolean result = _nc.deleteUser(TESTUSER);
+            assertTrue(result);
+        }
+    }
+
+    @Test
+    public void t19_01_testDeleteUserWithInvalidChar() {
+        System.out.println("deleteUserWithInvalidChar");
+        if (_nc != null)
+        {
+            boolean result = _nc.deleteUser(TESTUSER_WITH_INVALID_CHAR);
             assertTrue(result);
         }
     }